+++ packages/plugins/eslint-plugin-react-x/src/rules/no-duplicate-key.spec.ts
@@ -80,6 +80,21 @@ ruleTester.run(RULE_NAME, rule, {
         },
       ],
     },
+    {
+      code: /* tsx */ `
+        const App = () => {
+            return [1, 2, 3].map((item) => <div key={Date.now()}>{item}</div>)
+        };
+      `,
+      errors: [
+        {
+          messageId: "noDuplicateKey",
+          data: {
+            value: "{Date.now()}",
+          },
+        },
+      ],
+    },
   ],
   valid: [
     ...allValid,
@@ -95,11 +110,6 @@ ruleTester.run(RULE_NAME, rule, {
                   <div key="2">2</div>,
                   <div key="3">3</div>,
                ]
-      };
-    `,
-    /* tsx */ `
-      const App = () => {
-          return [1, 2, 3].map((item) => <div key={Math.random()}>{item}</div>)
       };
     `,
   ],